Retest on Windows.
Deployment fails to work under Windows. This commit is not the final fix. Also this commit adds some reorganizing of the script. I originally wanted a single CMakeLists.txt however it seems after all the workarounds needed it will be simpler for the user if I split the needed functionality in several different .cmake files.