BUG_FIX: When packaging on Windows the Qt6 build still couldn't execute winedeployqt.exe
In the script the mypath variable used forward slashes instead of backslashes and this ended in an error about an invalid path to the executable on the line that executes windeployqt.exe. Also the --no-angle command line option is no longer recognized in Qt6 so I made that Qt5 only.